Changed to flush apdu and marc log files a apdu is recived. Changed to remove blanks at end of commandline.
Added source of $HOME/.yazclientrc and ./yazclientrc on startup
added the folowing commands
  push_command   - adds the rest of the line to
  set_apdufile   - change apdu logfile from commandline
  set_marcdump   - change marcdump file form commandline
  set_cclfields  - change Ccl2rpn config file
  register_oid   - regristrate a private oid
Cleandup the tabcompletion and when it compleates on
filenames.
added completions for the follwing commands
  push_command
  format
  schema
  attributeset
  querytype

new pr. command completion is placed in tabcomplete.[ch]

Current problems:
  - the tabcompleation code is leaking - not mutch bot leaking.

Added emacs local versiables comment for tabs config to client.c