Refactor ICU WRT locale.
The locale parameter is no longer a parameter for icu_chain_xml_config
and locale is read as an attribute for top XML node. The icu_casemap
locale is no longer a member and a parater for icu_casemap_casemap
instead. This saves a copy of a locale in each casemap.