Do not handle EINTR in yaz_poll.
Change yaz_poll so that it does not loop around select/poll if
errno=EINTR is set. This sitution is handled in caller, such
as event_loop used by GFS which must take further action if EINTR
is set. If EINTR is handled by yaz_poll it does not get caught.