Fix compilation for mingw.
Two sources files failed to compile on GCC mingw (Minimalist GNU for
Windows). For oid_db.c the problem was that GCC reported 'initializer
element is not constant'. And the fix is to use initialize (use) at
run-time. For timing.c the problem that HAVE_SYS_TIME_H and WIN32 was
both defined - and the code was not prepared for that.