Introduce type for BER BOOLEAN/INTEGER .
The BER BOOLEAN is now the integral type Odr_bool which still happens
to be an 'int'. The BER INTEGER which used to be a plain 'int' is now
typedef'd to Odr_int. This is still an integral 'int', but it may be
changed to 64-bit via a one-liner in nmem.h (NMEM_64=1). The documentation,
ASN-1 compiler as well as programs has been updated to use the new types.