RPN to CQL conv may use USE string attributes (bug #2978).
In conversion from RPN to CQL (function cql_transform_rpn2cql_).. USE
string attributes are used as a fallback if there is no match in the
CQL transform (properties) lookup. Also change behavior so that
unsupported use attribute if there is no match .. Ie if there is a
numeric use attribute in a query and that is not matched anywhere.