ZOOM: resultsets not ref counted by connections
Experimental code and revert to old behavior of result sets.
ZOOM result sets are no longer ref counted by connections: it makes
it very difficult to work with multiple result sets within one
connection because connection would have to be destroyed before all
its result sets are destroyed. It will also makes use of more result
sets IDs on server than we need to  - because ALL result sets in use
will be still available (despite that ZOOM_resultset_destroy is
called). This commit reverts to existing behavior: a result set is
destroyed If ZOOM_resultset_destroy is used AND all operations on it
are completed.