SOLR support for ZOOM and yaz-client
The SOLR requests and responses (search only) are encoded in the
same structures as SRW/SRU, because they are semantically so equal.
It also means minimal change to existing software. However, at this
time decoding of a SOLR search response must be instructed by the client.
Had the SOLR Webservice used a namespace we might have a cleaner
model. For ZOOM, option "sru" may be set to "solr" in which case
ZOOM will contact a SOLR server with the URL with suffix "/select".
So far the following SRU parameters are mapped: query, startRecord,
maximumRecords.