Put yaz_log private data in non-static data
Move yaz_log private data, such as log file, file type, log
prefix from static to non-static (in a struct). This is to fix
a problem where a shared object will see the "old" initialized data
- not "current" data. For example, yaz_log_file would return stderr
handle in shared object program despite the fact that yaz_log_init_file
was called in mother program. The "private" data is combined in a
struct yaz_log_info to minimize number of globals (that really
are private data to yaz_log).