Avoid printf %.*s for arbitrary strings
Use wrbuf_write instead of wrbuf_printf in a few places for fixed
length buffers. Use wrbuf_puts instead of wrbuf_printf when printf
is not needed. There is a problem with printf("%.*", len, buf)
when buf is in different encoding from what is given by LANG
variable.. At least when readline is in use.