daemon: further tweaks to keepalive handling
SIGTERM: we wait until child terminates. This ensures that stop
script can safely assume that when moniotor process is gone, the
listening socket is closed.
SIGHUP: we send a signal to child process. Wait 2 seconds and stop
without waiting for child to terminate. Indeed it may still be running
serving remaining request. We have no way to know if the child has
really closed the listening socket within the two seconds we wait.