Decode SRU 2.0 responses
The official namespace for responses, according to
http://docs.oasis-open.org/search-ws/searchRetrieve/v1.0/cs01/schemas/sruResponse.xsd
appears to be http://docs.oasis-open.org/ns/search-ws/sruResponse .
nature.com's server, however, usee namespace
http://docs.oasis-open.org/ns/search-ws/sru-2-0-response
For this reason we are now using a glob-expr to match up the namespace.
ZOOM C and yaz-client updated.
Perhaps SRU 1.2 uses same namespace, I don't know. I can not find
a single real example showing the "real" namespace in the documentation
from OASIS. The namespace for requests is different from responses, but
this is not added to YAZ yet, when we don't encode them yet.