Attempt ISO2709 decoding of any octet aligned record
Until this point only certain OIDs would trigger ISO2709 decoding.
With this change all octet aligned records will be run though
the MARC parser. If it's not MARC, an error will only occur if it is
a known MARC OID. If not, all bets are off. It's probably not MARC
if it doesn't parse.