Partly revert UTF-8 records returned by yaz-ztest
Partly reverts commit 91ba2b5969d3de05dc2a2a127a1b17c7c9d07af7.
Not an error as such, but keep records as they are because yaz-ztest
returned records are used in pazpar2's test set. The records at 24
and later are UTF-8 encoded, though, to retain that test. And 50$a
has a capital o-slash in it.